+++ CHANGES.md
@@ -1,5 +1,11 @@
 [unreleased]
 
+#### 0.7.1 / 2021-03-16
+* add filter `gu_parse_api_branches`
+* add filter `gu_display_repos`
+* add filter `gu_running_git_servers`
+* add filter `gu_decode_response`
+
 #### 0.7.0 / 2021-03-15 🎂
 * add filter `gu_post_api_response_body`
 * add filter `gu_get_git_icon_data`